+++ jdem846/src/us/wthr/jdem846/DemConstants.java
@@ -25,12 +25,12 @@ public class DemConstants {
 	
 	public static final String APPLICATION_SPEC_VERSION = "0.1";
 	
-	public static final float EARTH_MEAN_RADIUS = 6371.0f;
-	public static final float MARS_MEAN_RADIUS = 3396.2f;
-	public static final float MOON_MEAN_RADIUS = 1737.1f;
+	public static final double EARTH_MEAN_RADIUS = 6371.0f;
+	public static final double MARS_MEAN_RADIUS = 3396.2f;
+	public static final double MOON_MEAN_RADIUS = 1737.1f;
 	
 	
-	public static final float ELEV_NO_DATA = -99999.99f;
+	public static final double ELEV_NO_DATA = -99999.99;
 	
 	
 	// Status Codes
